refactor(dronecan): extract DNA server into separate files
