+++ src/main/CMakeLists.txt
@@ -171,6 +171,8 @@ main_sources(COMMON_SRC
     drivers/dronecan/libcanard/canard.h
     drivers/dronecan/libcanard/canard_stm32_driver.h
     drivers/dronecan/dronecan.c
+    drivers/dronecan/dronecan_dna_server.c
+    drivers/dronecan/dronecan_dna_server.h
     drivers/dronecan/dronecan.h
      
     drivers/display.c
