+++ src/main/drivers/dronecan/dronecan_dna_server.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <stdint.h>
+#include "libcanard/canard.h"
+
+extern CanardInstance canard;
+
+void dronecanDnaHandleAllocation(CanardInstance *ins, CanardRxTransfer *transfer);
